+++ director/target/target.go
@@ -7,7 +7,9 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/greenbone/eulabeia/messages"
+	"github.com/greenbone/eulabeia/messages/cmds"
 	"github.com/greenbone/eulabeia/messages/handler"
+	"github.com/greenbone/eulabeia/messages/info"
 	"github.com/greenbone/eulabeia/models"
 	"github.com/greenbone/eulabeia/storage"
 )
@@ -16,20 +18,22 @@ type targetAggregate struct {
 	storage Storage
 }
 
-func (t targetAggregate) Create(c messages.Create) (*messages.Created, error) {
+func (t targetAggregate) Create(c cmds.Create) (*info.Created, error) {
 	target := models.Target{
 		ID: uuid.NewString(),
 	}
 	if err := t.storage.Put(target); err != nil {
 		return nil, err
 	}
-	return &messages.Created{
-		ID:      target.ID,
-		Message: messages.NewMessage("created.target", c.MessageID, c.GroupID),
+	return &info.Created{
+		Identifier: messages.Identifier{
+			Message: messages.NewMessage("created.target", c.MessageID, c.GroupID),
+			ID:      target.ID,
+		},
 	}, nil
 }
 
-func (t targetAggregate) Modify(m messages.Modify) (*messages.Modified, *messages.Failure, error) {
+func (t targetAggregate) Modify(m cmds.Modify) (*info.Modified, *info.Failure, error) {
 	var target *models.Target
 	target, err := t.storage.Get(m.ID)
 	if err != nil {
@@ -48,17 +52,19 @@ func (t targetAggregate) Modify(m messages.Modify) (*messages.Modified, *message
 		return nil, nil, err
 	}
 
-	return &messages.Modified{
-		ID:      m.ID,
-		Message: messages.NewMessage("modified.target", m.MessageID, m.GroupID),
+	return &info.Modified{
+		Identifier: messages.Identifier{
+			Message: messages.NewMessage("modified.target", m.MessageID, m.GroupID),
+			ID:      m.ID,
+		},
 	}, nil, nil
 
 }
-func (t targetAggregate) Get(g messages.Get) (interface{}, *messages.Failure, error) {
+func (t targetAggregate) Get(g cmds.Get) (messages.Event, *info.Failure, error) {
 	if target, err := t.storage.Get(g.ID); err != nil {
 		return nil, nil, err
 	} else if target == nil {
-		return nil, &messages.Failure{
+		return nil, &info.Failure{
 			Message: messages.NewMessage("failure.get.target", g.MessageID, g.GroupID),
 			Error:   fmt.Sprintf("%s not found.", g.ID),
 		}, nil
@@ -70,13 +76,15 @@ func (t targetAggregate) Get(g messages.Get) (interface{}, *messages.Failure, er
 	}
 }
 
-func (t targetAggregate) Delete(d messages.Delete) (*messages.Deleted, *messages.Failure, error) {
+func (t targetAggregate) Delete(d cmds.Delete) (*info.Deleted, *info.Failure, error) {
 	if err := t.storage.Delete(d.ID); err != nil {
-		return nil, messages.DeleteFailureResponse(d.Message, "target", d.ID), nil
+		return nil, info.DeleteFailureResponse(d.Message, "target", d.ID), nil
 	}
-	return &messages.Deleted{
-		Message: messages.NewMessage("deleted.target", d.MessageID, d.GroupID),
-		ID:      d.ID,
+	return &info.Deleted{
+		Identifier: messages.Identifier{
+			Message: messages.NewMessage("deleted.target", d.MessageID, d.GroupID),
+			ID:      d.ID,
+		},
 	}, nil, nil
 }